requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.9.0
pydantic-settings>=2.6.0
supabase>=2.10.0
twilio>=9.3.0
fpdf2>=2.8.0
python-dotenv>=1.0.0
